Started Game Screen, Added "Loading games...", fixed bug allowing users to add duplicate games, made names pretty on front

+++ Thatso/ProfileViewTableViewCell.h
@@ -12,5 +12,5 @@
 @property (nonatomic, strong) IBOutlet UIImageView *profilePicture;
 @property (nonatomic, strong) IBOutlet UILabel *nameLabel;
 
--(void)setColorScheme:(int) code;
+-(void)setColorScheme:(NSInteger) code;
 @end
